requirements.txt
pytorch-lightning==2.1.0
torch==2.1.2
torchvision==0.16.2
Pillow
open3d
h5py
pymcubes
fasteners
ftfy
regex
pytorch-wavelets
PyWavelets
spconv-cu120
numpy<2
setuptools==69.5.1
einops
timm
trimesh
seaborn
scikit-image